+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/components/organisms/trades/CloseTradeDialog.kt
@@ -2,7 +2,7 @@ package network.bisq.mobile.presentation.ui.components.organisms.trades
 
 import androidx.compose.runtime.Composable
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.ui.components.molecules.ConfirmationDialog
+import network.bisq.mobile.presentation.ui.components.molecules.dialog.ConfirmationDialog
 
 @Composable
 fun CloseTradeDialog(
